refactor(types): extract shared types for multi-consumer pages and fix duplications
- Create 9 shared type files in /types/ for genuinely multi-consumer interfaces
  (moderation, tickets, admin-users, payments, subscriptions, feedback, participants, payouts, disputes)
- Extend types/checkout.ts with AppliedDiscount interface (used by 4 checkout pages)
- Update ~30 dashboard/checkout/participant pages to import from shared types
- Extract duplicated FeatureItem component from ClassDetails + WebinarDetails into shared local file
- Fix import ordering in consultation checkout page (imports after type defs)
- Update moderation API route to use shared types with explicit date serialization
- Net reduction of ~650 lines of duplicated type definitions
