+++ .serena/memories/suggested_commands.md
@@ -1,0 +1,7 @@
+# Suggested Commands
+
+- `npm run test` — Run all tests
+- `npx tsc --noEmit` — Check TypeScript types
+- `npm run dev` — Start development server
+- `npx prisma generate` — Regenerate Prisma client
+- `git` — Standard git commands (Darwin/macOS)
