+++ app/api/staff/support-tickets/route.ts
@@ -15,6 +15,7 @@ import {
   Prisma,
 } from "@prisma/client";
 
+
 /**
  * GET /api/staff/support-tickets
  * List all support tickets with filters (staff/admin access)
