+++ app/dashboard/admin/feedback/page.tsx
@@ -50,36 +50,7 @@ import {
 } from "lucide-react";
 import { useToast } from "@/hooks/use-toast";
 import { FeedbackStatus } from "@prisma/client";
-
-interface FeedbackUser {
-  id: string;
-  name: string | null;
-  email: string | null;
-  image: string | null;
-  phone?: string | null;
-  createdAt?: string;
-}
-
-interface Feedback {
-  id: string;
-  title: string;
-  description: string;
-  rating: number | null;
-  category: string | null;
-  status: FeedbackStatus;
-  user: FeedbackUser;
-  createdAt: string;
-  updatedAt: string;
-}
-
-interface FeedbackCounts {
-  total: number;
-  pending: number;
-  acknowledged: number;
-  inProgress: number;
-  resolved: number;
-  closed: number;
-}
+import type { Feedback, FeedbackCounts } from "@/types/feedback";
 
 const STATUS_OPTIONS: { value: FeedbackStatus | "all"; label: string }[] = [
   { value: "all", label: "All Status" },
