+++ app/explore/programs/plans/classes/[classPlanId]/components/ClassDetails.tsx
@@ -28,6 +28,7 @@ import type {
   Topic,
 } from "@prisma/client";
 import { generateProgramImageUrl } from "@/app/explore/programs/utils";
+import { FeatureItem } from "@/app/explore/programs/plans/components/FeatureItem";
 
 type ClassSessionWithSchedule = PrismaClass & {
   appointments: (PrismaAppointment & {
@@ -66,23 +67,6 @@ const getBadgeVariant = (
   return "default";
 };
 
-type FeatureItemProps = {
-  icon: React.ReactNode;
-  label: string;
-  value: string | number | React.ReactNode;
-};
-
-const FeatureItem = ({ icon, label, value }: FeatureItemProps) => (
-  <div className="flex items-center gap-3 p-4 bg-zinc-50 rounded-xl">
-    <div className="w-10 h-10 rounded-lg bg-white border border-zinc-200 flex items-center justify-center text-zinc-600">
-      {icon}
-    </div>
-    <div>
-      <p className="text-xs text-zinc-500 uppercase tracking-wide">{label}</p>
-      <p className="text-sm font-semibold text-zinc-900">{value}</p>
-    </div>
-  </div>
-);
 
 interface ClassDetailsProps {
   readonly plan: ClassPlanDetailsData;
