+++ types/checkout.ts
@@ -1,5 +1,16 @@
 import { AppointmentsType, PaymentGateway } from "@prisma/client";
 
+/**
+ * Discount applied to a checkout session.
+ * Shared across all 4 checkout pages (consultation, subscription, webinar, class).
+ */
+export interface AppliedDiscount {
+  code: string;
+  discountType: "PERCENTAGE" | "FIXED_AMOUNT";
+  discountValue: number;
+  discountAmount?: number;
+}
+
 export interface CheckoutSession {
   id: string;
   slotId: string;
